define Appendable class for String and List
This required making Basics depend on List and String, so I needed to
make some changes to avoid module cycles.

Char, List, and String all needed to only depend on Haskell's prelude.
The only practical effect of this is that List.sortBy references the
Prelude.Ordering type instead of the Basics.Order type alias. This
could be resolved by having that type alias in its own module and
reexporting it in Basics.