update Parser/* to work on String type
The old versioned assumed that it would read from pinned memory, so it
only needed to hang on to two pointers. One for the current location,
and one for the end.

Now that a String may be moved by a GC, we need to keep track of the
underlying ByteArray# as well. This means doing an extra (+) each time
a Word8 is read.

I used ByteArray# directly so that this implementation can be used
without many changes if String someday switches from text-utf8 to a
(String ByteArray#) representation.